refactor: move modals to the dir
