+++ src/app/components/modals/category-confirm-modal/category-confirm-modal.ts
@@ -1,5 +1,5 @@
 import { Component, effect, input, output } from '@angular/core';
-import { Category } from '../../models/category.model';
+import { Category } from '../../../models/category.model';
 
 @Component({
   selector: 'app-category-confirm-modal',
