+++ src/app/components/modals/finish-modal/finish-modal.scss
