+++ src/app/components/modals/result-modal/result-modal.html
